T/requirements.txt
flask==3.0.0
flask-cors==4.0.0
google-api-python-client==2.100.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
requests==2.31.0
beautifulsoup4==4.12.2
openai==1.3.0
python-dotenv==1.0.0
email-validator==2.1.0
lxml==4.9.3
